add once flag to compilers

+++ docker/docker_compile.ps1
@@ -1,4 +1,7 @@
-param([switch]$o)
+param(
+    [switch]$o,
+    [switch]$once
+)
 
 $projectRoot = Split-Path -Parent $PSScriptRoot
 $dockerfile = Join-Path $PSScriptRoot "Dockerfile"
@@ -24,7 +27,9 @@ if ($LASTEXITCODE -ne 0) {
     exit 1
 }
 
-docker run --rm -v "${projectRoot}:/doc" oep-latex pdflatex -shell-escape -synctex=1 -output-directory=bin -jobname=oop-handbook main.tex
+if (-not $once) {
+    docker run --rm -v "${projectRoot}:/doc" oep-latex pdflatex -shell-escape -synctex=1 -output-directory=bin -jobname=oop-handbook main.tex
+}
 
 if ($LASTEXITCODE -eq 0) {
     Write-Host "Done: bin/oop-handbook.pdf"
